requirements.txt
Werkzeug==2.3.0
Flask-SQLAlchemy==3.0.2
Flask==2.3.1
Flask-Cors==3.0.10
flask-restx==1.1.0
itsdangerous==2.1.2
psycopg2-binary==2.9.6
bcrypt==4.0.1
pytest==7.3.1
pytest-cov==4.0.0
PyJWT==2.6.0